pyproject.toml
[project]
name = "cora"
version = "0.1.0"
description = "CORA stands for Cloud Operations Reasoning Agent. Cora helps developers and architects query their AWS architecture for resources and reason about the resource configurations."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.52",
    "fastapi>=0.104.0",
    "gradio>=4.0.0",
    "langfuse>=3.14.4",
    "openinference-instrumentation-smolagents>=0.1.22",
    "pyyaml>=6.0.0",
    "smolagents[telemetry]>=1.24.0",
    "textual>=8.0.0",
    "uvicorn>=0.24.0",
    "pyfiglet>=1.0.0",
    "aiosqlite>=0.22.1",
]

[dependency-groups]
dev = [
    "baml-py>=0.219.0",
    "ipykernel>=7.2.0",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
    "textual-dev>=1.8.0",
]
